pyproject.toml
[tool.poetry]
name = "aluguel_de_salas"
version = "0.1.0"
description = ""
authors = ["Felipe Almeida <felipeortalmeida@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
